backend/pyproject.toml
[project]
name = "vibecity-backend"
version = "0.1.0"
description = "VibeCity Real-time Backend (FastAPI)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "stripe>=8.1.0",
    "websockets>=12.0",
    "supabase>=2.3.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.7",
    "redis>=5.0.0",
    "python-dotenv>=1.0.1",
    "google-auth>=2.29.0",
    "requests>=2.31.0",
    "h3>=4.4.0",
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "opentelemetry-instrumentation-httpx>=0.47b0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "httpx>=0.26.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [".vercel_python_packages"]

[tool.ruff.lint]
select = ["E","F","I","B","UP"]
ignore = ["E501", "B008", "B904", "E402", "UP007", "F841", "B007", "UP042", "F401", "B017", "F821", "UP035", "I001"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
addopts = "-x -q"
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "network: marks tests requiring external network access",
]
